balance data type changed

+++ src/main/java/bank/Account.java
@@ -4,9 +4,9 @@ public class Account {
 
   private int id;
   private String type;
-  private int balance;
+  private double balance;
 
-  public Account(int id,String type,int balance)
+  public Account(int id,String type,double balance)
   {
     setId(id);
     setType(type);
@@ -30,11 +30,11 @@ public class Account {
     this.type = type;
   }
 
-  public int getBalance() {
+  public double getBalance() {
     return this.balance;
   }
 
-  public void setBalance(int balance) {
+  public void setBalance(double balance) {
     this.balance = balance;
   }
